// packages/angular/entity/entity-schema/types.ts
export enum EntityCapacity {
  Dimension = 'Dimension',
  Hierarchy = 'Hierarchy',
  Measure = 'Measure',
  Indicator = 'Indicator',
  Calculation = 'Calculation',
  Parameter = 'Parameter',
}

export enum EntitySchemaType {
  Entity = 'Entity',
  Dimension = 'Dimension',
  Hierarchy = 'Hierarchy',
  Level = 'Level',
  IMeasure = 'IMeasure',
  Member = 'Member',
  Parameters = 'Parameters',
  Parameter = 'Parameter',
  Properties = 'Properties',
  Field = 'Field'
}